fix: resolve bugs B01-B30, L01-L08 from bugs.md audit
Critical fixes:
- B01: Mutex poisoning panics in agent rate limiter (unwrap_or_else)
- B02: HashMap unwrap in orchestrator DAG resolution (guard check)
- B03: Lua VM init panic (log error before panic)
- B04: HTTP client creation panics (fallback client)
- B05: Aho-Corasick empty pattern panic (Option wrapper)
- B06: Kafka response length from untrusted data (validate non-negative)
- B07: LDAP length parsing out-of-bounds (bounds checks)
- B08: NTP response slicing without length check (validate len >= 48)
- B09: SMB response buffer allocation overflow (checked_add + bounds)
- B10: Kafka topic/key/value length wrapping (validate before cast)
- B11: PostgreSQL query length truncation (try_into)

High-severity fixes:
- B13: Nested locking deadlock risk (scope scripts_path lock)
- B14: TCP connections without timeouts (8 locations, add timeout())
- B15: SMB_SESSIONS never cleaned (bounded session cache)
- B16: shortport unbounded port range expansion (skip invalid ranges)
- B17: PostgreSQL unbounded response accumulation (max 16MB)
- B18: TUI scrollable.rs underflow (saturating_sub)
- B19: TUI selector.rs underflow (saturating_sub)
- B20: Blocking I/O in async contexts (18 locations, tokio::fs)

Medium-severity fixes:
- B21: Silently dropped socket timeout settings (log errors)
- B22: Silently dropped network write failures (log errors)
- B23: Strict mode flag set failure (log errors)
- B27: Rate limiter overflow (saturating arithmetic)
- B28: Trend analysis integer overflow (i32 -> i64)

Low-severity fixes:
- L02: Kafka message length truncation (validate before cast)
- L03: Shortport port number wrapping (bounds check)
- L04: Recon retry delay exponential overflow (saturating_pow + cap)
- L05: Finding banner truncation mismatch (compare trimmed len)
- L08: Unnecessary clone in proxy intercept (iter().cloned())

+++ crates/eggsec-output/src/trend.rs
@@ -176,19 +176,19 @@ impl TrendAnalyzer {
         let mut sorted_results: Vec<_> = self.results.iter().collect();
         sorted_results.sort_by(|a, b| a.1.timestamp.cmp(&b.1.timestamp));
 
-        let critical_trend: Vec<i32> = sorted_results
+        let critical_trend: Vec<i64> = sorted_results
             .windows(2)
-            .map(|w| w[1].1.summary.critical as i32 - w[0].1.summary.critical as i32)
+            .map(|w| w[1].1.summary.critical as i64 - w[0].1.summary.critical as i64)
             .collect();
 
-        let high_trend: Vec<i32> = sorted_results
+        let high_trend: Vec<i64> = sorted_results
             .windows(2)
-            .map(|w| w[1].1.summary.high as i32 - w[0].1.summary.high as i32)
+            .map(|w| w[1].1.summary.high as i64 - w[0].1.summary.high as i64)
             .collect();
 
-        let medium_trend: Vec<i32> = sorted_results
+        let medium_trend: Vec<i64> = sorted_results
             .windows(2)
-            .map(|w| w[1].1.summary.medium as i32 - w[0].1.summary.medium as i32)
+            .map(|w| w[1].1.summary.medium as i64 - w[0].1.summary.medium as i64)
             .collect();
 
         let total_duration: u64 = sorted_results
@@ -241,9 +241,9 @@ impl TrendAnalyzer {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct TrendAnalysis {
     pub direction: TrendDirection,
-    pub critical_trend: Vec<i32>,
-    pub high_trend: Vec<i32>,
-    pub medium_trend: Vec<i32>,
+    pub critical_trend: Vec<i64>,
+    pub high_trend: Vec<i64>,
+    pub medium_trend: Vec<i64>,
     pub average_scan_time_ms: u64,
 }
